+++ src/app/signup/page.tsx
@@ -2,8 +2,8 @@
 
 import Link from 'next/link';
 
-import LeftSection from '@/components/LeftSection';
-import LogoMobile from '@/components/LogoMobile';
+import LeftSection from '@/components/left-section-login';
+import LogoMobile from '@/components/logo-mobile';
 import { Button } from '@/components/ui/button';
 import { Checkbox } from '@/components/ui/checkbox';
 import { Input } from '@/components/ui/input';
