+++ src/components/logo-mobile/index.tsx
