+++ src/lib/axios.ts
@@ -22,7 +22,6 @@ if (typeof window !== 'undefined') {
     (error) => {
       if (error.response?.status === 401) {
         localStorage.removeItem('token');
-        window.location.href = '/login';
       }
       return Promise.reject(new Error(error.message));
     }
